requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
dotenv==0.9.9
    # via -r requirements.in
idna==3.10
    # via requests
numpy==2.2.6
    # via pandas
pandas==2.3.3
    # via -r requirements.in
pydantic==2.12.2
    # via -r requirements.in
pydantic-core==2.41.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via dotenv
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
structlog==25.4.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   structlog
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
